MAP!!!
Made the map. I don't know if this commit will work but I have a backup.
Put everything together, navmesh is set up, prefabs created, textures
downloaded, wrists slit.

Everything should work out of the box and is in the scene ss1. There is
a Heavy set up in the scene already.