Pickups, shooting fixes for buffs
Created Health and Ammo pickups which give the player character that
touches them health or ammo. The pickup then disappears for a set time
and reappears to be picked up again.

Fixed shooting for Heavy and Ninja. Problem was caused by adding a
damage modifier to the projectile instance. Unity is funny with the way
it retains instances of prefabs.