SFX, HUD and extras
Updated the navmesh (testing, it may be broken) and changed the map a
little.

Made sure that the HUD adds the correct player portrait, and added the
scopecam overlay for the sniper.

Added all the SFX for the characters. We can change this again later but
the sound pool is populated for the main characters.