Game Obera, Scene, HUD, some tweaks
Updated Navmesh and map, though things might go a little wrong with
this.

Game over screen works. Players can also quit the game, and when they
all die, they go back to the main menu. Works quite well.

The HUD was also updated to show specials, but tweaks need to be made to
this. Character portraits are also missing again.