UI and Music
Updated the game overlay to show the total game time in place of the
score. We can use this for the future now.

Updated Combat music volume, since it was much too soft earlier, and
delayed it's start so we hear the diminished notes the first time round.

Changed something in the round manager so we actually can get more
enemies spawning later on. It was stuck at 15 but shouldn't be anymore.